lol2youtube/requirements.txt
asgiref==3.4.1
Django==3.2.8
djangorestframework==3.13.1
gunicorn==20.1.0
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
